move all failing v10 tests to repositories:: modules
