+++ src/lib/src/core/v0_10_0/index/commit_entry_writer.rs
@@ -1525,53 +1525,3 @@ impl CommitEntryWriter {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-
-    use crate::core::v0_10_0::index::CommitEntryReader;
-    use crate::error::OxenError;
-    use crate::repositories;
-    use crate::test;
-    use std::path::PathBuf;
-
-    #[tokio::test]
-    async fn test_merkle_two_files_same_hash() -> Result<(), OxenError> {
-        test::run_empty_local_repo_test_async(|local_repo| async move {
-            let p1 = "hi.txt";
-            let p2 = "bye.txt";
-            let path_1 = local_repo.path.join(p1);
-            let path_2 = local_repo.path.join(p2);
-
-            let common_contents = "the same file";
-
-            test::write_txt_file_to_path(&path_1, common_contents)?;
-            test::write_txt_file_to_path(&path_2, common_contents)?;
-
-            repositories::add(&local_repo, &path_1)?;
-            repositories::add(&local_repo, &path_2)?;
-
-            let status = repositories::status(&local_repo)?;
-
-            log::debug!("staged files here are {:?}", status.staged_files);
-
-            assert_eq!(status.staged_files.len(), 2);
-
-            assert!(status.staged_files.contains_key(&PathBuf::from(p1)));
-            assert!(status.staged_files.contains_key(&PathBuf::from(p2)));
-
-            let commit = repositories::commit(&local_repo, "add two files")?;
-
-            let commit_entry_reader = CommitEntryReader::new(&local_repo, &commit)?;
-
-            // List all the files
-            let _files = commit_entry_reader.list_entries()?;
-
-            assert!(commit_entry_reader.has_file(&PathBuf::from(p1)));
-            assert!(commit_entry_reader.has_file(&PathBuf::from(p2)));
-
-            Ok(())
-        })
-        .await
-    }
-}
