+++ src/lib/src/core/v0_19_0/workspaces/data_frames.rs
@@ -36,8 +36,6 @@ pub fn get_queryable_data_frame_workspace(
 ) -> Result<Workspace, OxenError> {
     let path = path.as_ref();
 
-    let commit_merkle_tree = CommitMerkleTree::from_path(repo, commit, path, true)?;
-
     let file_node = repositories::tree::get_file_by_path(repo, commit, path)?
         .ok_or(OxenError::path_does_not_exist(path))?;
     if file_node.data_type != EntryDataType::Tabular {
