+++ src/lib/src/test.rs
@@ -989,6 +989,45 @@ where
     Ok(())
 }
 
+/// Run a test on a repo with a bunch of files
+pub async fn run_training_data_repo_test_fully_committed_async_min_version<T, Fut>(
+    version: MinOxenVersion,
+    test: T,
+) -> Result<(), OxenError>
+where
+    T: FnOnce(LocalRepository) -> Fut,
+    Fut: Future<Output = Result<(), OxenError>>,
+{
+    init_test_env();
+    log::info!("<<<<< run_training_data_repo_test_fully_committed_async_min_version start");
+    let repo_dir = create_repo_dir(test_run_dir())?;
+    let repo = repositories::init::init_with_version(&repo_dir, version)?;
+
+    // Write all the files
+    populate_dir_with_training_data(&repo_dir)?;
+    // Add all the files
+    repositories::add(&repo, &repo.path)?;
+    log::debug!("about to commit this repo");
+    repositories::commit(&repo, "adding all data baby")?;
+    log::debug!("successfully committed the repo");
+    // Run test to see if it panic'd
+    log::info!(">>>>> run_training_data_repo_test_fully_committed_async running test");
+    let result = match test(repo).await {
+        Ok(_) => true,
+        Err(err) => {
+            eprintln!("Error running test. Err: {err}");
+            false
+        }
+    };
+
+    // Remove repo dir
+    util::fs::remove_dir_all(&repo_dir)?;
+
+    // Assert everything okay after we cleanup the repo dir
+    assert!(result);
+    Ok(())
+}
+
 fn create_bounding_box_csv(repo_path: &Path) -> Result<(), OxenError> {
     let dir = repo_path.join("annotations").join("train");
     // Create dir
